refactor(widget): split LogViewer into modular structure
Split the 2168-line log_viewer.rs into organized modules:

- types.rs: LogLevel, TimestampFormat
- entry.rs: LogEntry, SearchMatch
- filter.rs: LogFilter with matching logic
- parser.rs: LogParser with JSON and standard format parsing
- mod.rs: LogViewer widget

This improves code organization and maintainability while preserving
all functionality and exports.
